Gate v.2.0

+++ app/Http/Controllers/RoleController.php
@@ -5,6 +5,7 @@ namespace App\Http\Controllers;
 use App\Permisos\Models\Permission;
 use App\Permisos\Models\Role;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Gate;
 
 class RoleController extends Controller
 {
@@ -15,6 +16,7 @@ class RoleController extends Controller
      */
     public function index()
     {
+        Gate::authorize("haveaccess", 'role.index');
         $roles = Role::orderBy('id', 'Desc')->paginate(2);
 
         return view('role.index', compact('roles'));
@@ -27,6 +29,7 @@ class RoleController extends Controller
      */
     public function create()
     {
+        Gate::authorize("haveaccess", 'role.create');
         $permissions = Permission::get();
 
         return view('role.create', compact('permissions'));
@@ -40,6 +43,7 @@ class RoleController extends Controller
      */
     public function store(Request $request)
     {
+        Gate::authorize("haveaccess", 'role.create');
         $request->validate([
             'name' => 'required|max:50|unique:roles,name',
             'slug' => 'required|max:50|unique:roles,slug',
@@ -64,6 +68,7 @@ class RoleController extends Controller
      */
     public function show(Role $role)
     {
+        Gate::authorize("haveaccess", 'role.show');
         $permissions_role = [];
 
         foreach($role->permissions as $p){
@@ -83,6 +88,7 @@ class RoleController extends Controller
      */
     public function edit(Role $role)
     {
+        Gate::authorize("haveaccess", 'role.edit');
         $permission_role = [];
 
         foreach($role->permissions as $p){
@@ -103,6 +109,7 @@ class RoleController extends Controller
      */
     public function update(Request $request, Role $role)
     {
+        Gate::authorize("haveaccess", 'role.edit');
         $request->validate([
             'name' => 'required|max:50|unique:roles,name,'.$role->id,
             'slug' => 'required|max:50|unique:roles,slug,'.$role->id,
@@ -126,6 +133,7 @@ class RoleController extends Controller
      */
     public function destroy(Role $role)
     {
+        Gate::authorize("haveaccess", 'role.destroy');
         $role->delete();
 
         return redirect()->route('role.index')->with('status_success', 'Role successfully removed');
